Replace BigDecimal with BigInteger to simplify things

+++ app/src/androidTest/java/com/concordium/wallet/data/room/AccountDaoUnitTest.kt
@@ -7,7 +7,7 @@ import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.concordium.wallet.data.model.*
 import com.concordium.wallet.getOrAwaitValue
-import com.concordium.wallet.util.toBigDecimal
+import com.concordium.wallet.util.toBigInteger
 import kotlinx.coroutines.runBlocking
 import org.junit.After
 import org.junit.Assert.assertEquals
@@ -16,7 +16,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.io.IOException
-import java.math.BigDecimal
+import java.math.BigInteger
 
 @RunWith(AndroidJUnit4::class)
 class AccountDaoUnitTest {
@@ -57,16 +57,16 @@ class AccountDaoUnitTest {
             "",
             revealedAttributes,
             CredentialWrapper(RawJson("{}"), 1),
-            "115792089237316195423570985008687907853269984665640564039457584007913129639935".toBigDecimal(),
-            BigDecimal.ZERO,
-            BigDecimal.ZERO,
-            BigDecimal.ZERO,
-            BigDecimal.ZERO,
+            "115792089237316195423570985008687907853269984665640564039457584007913129639935".toBigInteger(),
+            BigInteger.ZERO,
+            BigInteger.ZERO,
+            BigInteger.ZERO,
+            BigInteger.ZERO,
             null,
             null,
             ShieldedAccountEncryptionStatus.ENCRYPTED,
-            BigDecimal.ZERO,
-            BigDecimal.ZERO,
+            BigInteger.ZERO,
+            BigInteger.ZERO,
             false,
             null,
             null,
