+++ app/src/main/java/com/concordium/wallet/ui/account/accountdetails/AccountReleaseScheduleActivity.kt
@@ -18,7 +18,7 @@ import com.concordium.wallet.databinding.AccountReleaseScheduleItemBinding
 import com.concordium.wallet.databinding.AccountReleaseScheduleTransactionItemBinding
 import com.concordium.wallet.databinding.ActivityAccountReleaseScheduleBinding
 import com.concordium.wallet.ui.base.BaseActivity
-import com.concordium.wallet.util.toBigDecimal
+import com.concordium.wallet.util.toBigInteger
 import java.text.DateFormat
 import java.util.*
 
@@ -90,7 +90,7 @@ class AccountReleaseScheduleActivity : BaseActivity() {
         })
         viewModel.scheduledReleasesLiveData.observe(this, Observer<List<Schedule>> { list ->
 
-            binding.accountReleaseScheduleLockedAmount.text = CurrencyUtil.formatGTU(viewModel.account.finalizedAccountReleaseSchedule?.total.toBigDecimal(), true)
+            binding.accountReleaseScheduleLockedAmount.text = CurrencyUtil.formatGTU(viewModel.account.finalizedAccountReleaseSchedule?.total.toBigInteger(), true)
 
             binding.accountReleaseScheduleList.removeAllViews()
             val dateFormat = DateFormat.getDateTimeInstance(DateFormat.SHORT, DateFormat.SHORT, Locale.getDefault())
@@ -113,7 +113,7 @@ class AccountReleaseScheduleActivity : BaseActivity() {
                     }
                 }
 
-                view.amount.text = CurrencyUtil.formatGTU(release.amount.toBigDecimal(), true)
+                view.amount.text = CurrencyUtil.formatGTU(release.amount.toBigInteger(), true)
                 binding.accountReleaseScheduleList.addView(view.root)
             }
         })
